Checkstyle Revisions (#50)
* Checkstyle Revisions

Testing removing some check styles for easier collaboration.

-Removed NoLineWrap (allows flexibility in import statements)
- Removed NeedBraces (allows single-line if and loops without braces)
-Removed AvoidStarImport (allows wildcard imports for convenience)
-Removed MissingSwitchDefault (default case in switch statements no longer mandatory)
-Removed ModifierOrder (modifiers can be in any order)
- Relaxed Naming Rules (still enforces readability but allows some flexibility)
- Removed AnnotationLocation (allows annotations anywhere without warnings)

* error upon proposal, fix.

* Update checkstyle_checks.xml (fix 2)

Removed the Tab Checker (doing a debug process this file to root down the errors)

* fix 3 (Read)

Changed some of the checkstyle so it's more lenient with the naming conventions. Original didn't allow the naming of files to be like NetworkAPI rather it had to be networkapi.

* fix 4

Naming convention fix.

* fix 5

Remove tab enforcement so it accepts tabs or spaces